Rename AliasSpec to TypeAliasSpec
